# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FinCommand Executive Treasury API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List bills with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, on_hold, partial, paid, overdue]
        - name: userId
          in: query
          schema:
            type: string
        - name: vendorId
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: object
                properties:
                  bills:
                    type: array
                    items:
                      $ref: "#/components/schemas/Bill"
                  total:
                    type: number
                required: [bills, total]
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a new bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillInput"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get bill details with comments and audit trail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bill details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillDetail"
    patch:
      operationId: updateBill
      tags: [bills]
      summary: Update bill fields
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillInput"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}/approve:
    post:
      operationId: approveBill
      tags: [bills]
      summary: Approve a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                approvedAmount:
                  type: number
      responses:
        "200":
          description: Approved bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}/reject:
    post:
      operationId: rejectBill
      tags: [bills]
      summary: Reject a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
      responses:
        "200":
          description: Rejected bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}/hold:
    post:
      operationId: holdBill
      tags: [bills]
      summary: Put a bill on hold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
                rescheduleDate:
                  type: string
      responses:
        "200":
          description: Bill on hold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}/partial-approve:
    post:
      operationId: partialApproveBill
      tags: [bills]
      summary: Partially approve a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approvedAmount:
                  type: number
                comment:
                  type: string
              required: [approvedAmount]
      responses:
        "200":
          description: Partially approved bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}/escalate:
    post:
      operationId: escalateBill
      tags: [bills]
      summary: Escalate an overdue bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment:
                  type: string
      responses:
        "200":
          description: Escalated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /bills/{id}/comments:
    get:
      operationId: getBillComments
      tags: [comments]
      summary: Get comments for a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bill comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  comments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Comment"
                required: [comments]
    post:
      operationId: addBillComment
      tags: [comments]
      summary: Add a comment to a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                authorId:
                  type: string
              required: [text, authorId]
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /bills/{id}/audit:
    get:
      operationId: getBillAudit
      tags: [audit]
      summary: Get audit trail for a bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audit trail
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEntry"
                required: [entries]

  # ─── VENDORS ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Bill:
      type: object
      properties:
        id:
          type: string
        vendorId:
          type: string
        vendorName:
          type: string
        description:
          type: string
        amount:
          type: number
        approvedAmount:
          type: number
        paidAmount:
          type: number
        outstandingBalance:
          type: number
        scheduledDate:
          type: string
        dueDate:
          type: string
        walletId:
          type: string
        walletName:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [pending, approved, rejected, on_hold, partial, paid, overdue]
        createdBy:
          type: string
        createdByName:
          type: string
        hasAttachment:
          type: boolean
        overdueDays:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - vendorId
        - vendorName
        - description
        - amount
        - paidAmount
        - outstandingBalance
        - scheduledDate
        - priority
        - status
        - createdBy
        - createdByName
        - createdAt
    BillDetail:
      allOf:
        - $ref: "#/components/schemas/Bill"
        - type: object
          properties:
            comments:
              type: array
              items:
                $ref: "#/components/schemas/Comment"
            auditTrail:
              type: array
              items:
                $ref: "#/components/schemas/AuditEntry"
            vendor:
              $ref: "#/components/schemas/Vendor"
          required: [comments, auditTrail]
    CreateBillInput:
      type: object
      properties:
        vendorId:
          type: string
        description:
          type: string
        amount:
          type: number
        scheduledDate:
          type: string
        dueDate:
          type: string
        walletId:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        createdBy:
          type: string
        hasAttachment:
          type: boolean
      required: [vendorId, description, amount, scheduledDate, priority, createdBy]
    UpdateBillInput:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        scheduledDate:
          type: string
        dueDate:
          type: string
        walletId:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
    Vendor:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        totalBilled:
          type: number
        totalPaid:
          type: number
        outstandingBalance:
          type: number
        lastPaymentDate:
          type: string
        createdAt:
          type: string
      required: [id, name, totalBilled, totalPaid, outstandingBalance, createdAt]
...
    Comment:
      type: object
      properties:
        id:
          type: string
        billId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        authorRole:
          type: string
        text:
          type: string
        createdAt:
          type: string
      required: [id, billId, authorId, authorName, authorRole, text, createdAt]
    AuditEntry:
      type: object
      properties:
        id:
          type: string
        billId:
          type: string
        userId:
          type: string
        userName:
          type: string
        action:
          type: string
        details:
          type: string
        oldValue:
          type: string
        newValue:
          type: string
        createdAt:
          type: string
      required: [id, userId, userName, action, createdAt]
...